Changes in JSON data structure, how plugin handles and sends data to frontend, how frontend manages data
The JSON data structure now has general info, meta data and translations, there are not translations directly into the data strucure, this to retain more info. This way the data now is handled in a slightly complicate way, also inside the rest APIS. This way the frontend will handle translations and meta data and the backend (server code) will update this data.