Major updates, bug fixing, ui/ux improvements, security improvements, readme updates, refactoring etc.
This new version of the plugin will include more features, updates, bug fixing and code refactoring best practices, the UI/UX is improved in React Typescript and CSS, more Types in Typescript are added, better handled and structured. A new data structure is used, now we do not have the old data structure where translations are added directly, but we have a more secure and structured one. Now we have general info, meta data and translations inside. Ids now are added to translations. The code is now more secure and robust, the frontend manages the updates, remove and the getting of the data better, also thanks to the addition of the IDs in the data structure. I added pagination, entries per page functionalities, entries found and entries searched, a search bar with a clean search button. The code for these functionalities is double reviewed. More functionalities, fixes and updates will come in the future