Major updates: ngettext/_n() strings translation. Plural/String translations. Bug fixing. UI/UX improvements
This plugin finally introduced translations for strings that can be plural and singular inserted with the wordpress function _n(). This function, based on how many elements there are in an array, will display the singular or plural version of the string. So, now this plugin will also use the ngettext filter to translate those strings using a custom function inside the columns where you insert the string to translate and the translation (the custom function is: %_ctp_n(singular: 'Singular string', plural: 'Plural string' and is well explained on the readme file and on the settings page of the plugin in the admin area.). Bug fixing and other issues were solved on the backend and on the frontend area. For the frontend pagination, search and adding rows related bugs were solved, on the backend area, the translations now were applied only on the backend. Plus, on the settings page where you use the plugin, accordions were added to display information on the plugin and how to use it.